chore: move cac instance to its own file
